Create handlers package and move status handler into it
