+++ healthcheck/main.go
@@ -1,26 +1,15 @@
 package main
 
 import (
+	"github.com/tsuru/nginx-operator/healthcheck/handlers"
 	"log"
 	"net/http"
-	"strings"
 )
 
 const listen = ":59999"
 
-func statusHandler(w http.ResponseWriter, r *http.Request) {
-	ports := r.URL.Query()["ports"]
-
-	if len(ports) < 1 {
-		w.WriteHeader(http.StatusBadRequest)
-	} else {
-		ports = strings.Split(ports[0], ",")
-		w.WriteHeader(http.StatusOK)
-	}
-}
-
 func main() {
-	http.HandleFunc("/status", statusHandler)
+	http.HandleFunc("/status", handlers.StatusHandler)
 
 	log.Fatal(http.ListenAndServe(listen, nil))
 }
